# string type arrays handling dates